refactor save_vocabulary
