+++ src/transformers/models/bert/tokenization_bert_fast.py
@@ -113,9 +113,6 @@ class BertTokenizerFast(PreTrainedTokenizerFast):
             self.backend_tokenizer.normalizer = normalizer_class(**normalizer_state)
 
 
-    def save_vocabulary(self, save_directory: str, filename_prefix: Optional[str] = None) -> Tuple[str]:
-        files = self._tokenizer.model.save(save_directory, name=filename_prefix)
-        return tuple(files)
 
 
 __all__ = ["BertTokenizerFast"]
