+++ src/transformers/models/convbert/tokenization_convbert_fast.py
@@ -139,9 +139,6 @@ class ConvBertTokenizerFast(PreTrainedTokenizerFast):
         return output
 
 
-    def save_vocabulary(self, save_directory: str, filename_prefix: Optional[str] = None) -> Tuple[str]:
-        files = self._tokenizer.model.save(save_directory, name=filename_prefix)
-        return tuple(files)
 
 
 __all__ = ["ConvBertTokenizerFast"]
