+++ src/transformers/models/deprecated/realm/tokenization_realm_fast.py
@@ -190,9 +190,6 @@ class RealmTokenizerFast(PreTrainedTokenizerFast):
 
         return BatchEncoding(output_data, tensor_type=return_tensors)
 
-    def save_vocabulary(self, save_directory: str, filename_prefix: Optional[str] = None) -> Tuple[str]:
-        files = self._tokenizer.model.save(save_directory, name=filename_prefix)
-        return tuple(files)
 
 
 __all__ = ["RealmTokenizerFast"]
