+++ src/transformers/models/distilbert/tokenization_distilbert_fast.py
@@ -115,9 +115,6 @@ class DistilBertTokenizerFast(PreTrainedTokenizerFast):
 
 
     # Copied from transformers.models.bert.tokenization_bert_fast.BertTokenizerFast.save_vocabulary
-    def save_vocabulary(self, save_directory: str, filename_prefix: Optional[str] = None) -> Tuple[str]:
-        files = self._tokenizer.model.save(save_directory, name=filename_prefix)
-        return tuple(files)
 
 
 __all__ = ["DistilBertTokenizerFast"]
