+++ src/transformers/models/electra/tokenization_electra_fast.py
@@ -110,9 +110,6 @@ class ElectraTokenizerFast(PreTrainedTokenizerFast):
             self.backend_tokenizer.normalizer = normalizer_class(**normalizer_state)
 
 
-    def save_vocabulary(self, save_directory: str, filename_prefix: Optional[str] = None) -> Tuple[str]:
-        files = self._tokenizer.model.save(save_directory, name=filename_prefix)
-        return tuple(files)
 
 
 __all__ = ["ElectraTokenizerFast"]
