+++ src/transformers/models/gpt2/tokenization_gpt2_fast.py
@@ -125,9 +125,6 @@ class GPT2TokenizerFast(PreTrainedTokenizerFast):
 
         return super()._encode_plus(*args, **kwargs)
 
-    def save_vocabulary(self, save_directory: str, filename_prefix: Optional[str] = None) -> Tuple[str]:
-        files = self._tokenizer.model.save(save_directory, name=filename_prefix)
-        return tuple(files)
 
 
 __all__ = ["GPT2TokenizerFast"]
