+++ src/transformers/models/gpt_neox/tokenization_gpt_neox_fast.py
@@ -159,9 +159,6 @@ class GPTNeoXTokenizerFast(PreTrainedTokenizerFast):
             + eos_token_id
         )
 
-    def save_vocabulary(self, save_directory: str, filename_prefix: Optional[str] = None) -> Tuple[str]:
-        files = self._tokenizer.model.save(save_directory, name=filename_prefix)
-        return tuple(files)
 
 
 __all__ = ["GPTNeoXTokenizerFast"]
