+++ src/transformers/models/layoutlmv3/tokenization_layoutlmv3_fast.py
@@ -812,9 +812,6 @@ class LayoutLMv3TokenizerFast(PreTrainedTokenizerFast):
         return encoded_inputs
 
     # Copied from transformers.models.layoutlmv2.tokenization_layoutlmv2_fast.LayoutLMv2TokenizerFast.save_vocabulary
-    def save_vocabulary(self, save_directory: str, filename_prefix: Optional[str] = None) -> Tuple[str]:
-        files = self._tokenizer.model.save(save_directory, name=filename_prefix)
-        return tuple(files)
 
     def build_inputs_with_special_tokens(self, token_ids_0, token_ids_1=None):
         output = [self.bos_token_id] + token_ids_0 + [self.eos_token_id]
