+++ src/transformers/models/markuplm/tokenization_markuplm_fast.py
@@ -876,9 +876,6 @@ class MarkupLMTokenizerFast(PreTrainedTokenizerFast):
 
         return encoded_inputs
 
-    def save_vocabulary(self, save_directory: str, filename_prefix: Optional[str] = None) -> Tuple[str]:
-        files = self._tokenizer.model.save(save_directory, name=filename_prefix)
-        return tuple(files)
 
 
 __all__ = ["MarkupLMTokenizerFast"]
