+++ src/transformers/models/openai/tokenization_openai_fast.py
@@ -58,9 +58,6 @@ class OpenAIGPTTokenizerFast(PreTrainedTokenizerFast):
     def do_lower_case(self):
         return True
 
-    def save_vocabulary(self, save_directory: str, filename_prefix: Optional[str] = None) -> Tuple[str]:
-        files = self._tokenizer.model.save(save_directory, name=filename_prefix)
-        return tuple(files)
 
 
 __all__ = ["OpenAIGPTTokenizerFast"]
