+++ src/transformers/models/splinter/tokenization_splinter_fast.py
@@ -184,9 +184,6 @@ class SplinterTokenizerFast(PreTrainedTokenizerFast):
             # Input is context-then-question
             return len(cls + token_ids_0 + sep) * [0] + len(token_ids_1 + question_suffix + sep) * [1]
 
-    def save_vocabulary(self, save_directory: str, filename_prefix: Optional[str] = None) -> Tuple[str]:
-        files = self._tokenizer.model.save(save_directory, name=filename_prefix)
-        return tuple(files)
 
 
 __all__ = ["SplinterTokenizerFast"]
