+++ src/transformers/models/squeezebert/tokenization_squeezebert_fast.py
@@ -114,9 +114,6 @@ class SqueezeBertTokenizerFast(PreTrainedTokenizerFast):
             self.backend_tokenizer.normalizer = normalizer_class(**normalizer_state)
 
 
-    def save_vocabulary(self, save_directory: str, filename_prefix: Optional[str] = None) -> Tuple[str]:
-        files = self._tokenizer.model.save(save_directory, name=filename_prefix)
-        return tuple(files)
 
 
 __all__ = ["SqueezeBertTokenizerFast"]
